requirements.txt
flatdict==1.2.0
pyyaml==3.11